aulas do dia 23 e 25 de Agosto e desafio do dia 19 de agosto

+++ backend/src/controllers/HelpersController.ts
@@ -1,12 +1,20 @@
 import { Request, Response } from 'express'
 
+import FoodTruck from '../models/FoodTruck'
 import User from '../models/User'
 
 export default {
 
     async resetUser(req: Request, res: Response) {
         const { instagram } = req.query
+        const user = await User.findOne({instagram: instagram})
+
+        if (!user)
+            return res.status(204).end()
+
+        await FoodTruck.deleteMany({suggested_by: user._id})
         await User.deleteMany({instagram: instagram})
-        res.status(204).end()        
+
+        return res.status(204).end()        
     }
 }
